multiple os and latest pushes
Operating System Matrix:
We're using a matrix strategy to run the job on different operating systems: Ubuntu, Windows, and macOS.
Only Trigger on Push to Main Branch:
The workflow will only be triggered when pushing changes to the main branch.
Check for Changes:
We use a script to determine which files have changed since the last commit. This will provide us with a list of changed files.